requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
tavily-python
langchain
langchain-groq
langgraph
openai